# backend/docs/api/property-management-openapi.yaml
openapi: 3.0.3
info:
  title: PropMetrik Property Management API
  description: |
    Property Management API for PropMetrik platform.
    
    ## Features
    - **Property Management**: Properties, units, amenities, and listings
    - **Tenancy Management**: Lease lifecycle, tenant onboarding, renewals
    - **Payment Processing**: Rent collection via Paystack (cards, mobile money)
    - **Maintenance**: Work order submission and status tracking
    - **Lease Templates**: Template CRUD, variable substitution, PDF generation
    - **Tenant Portal**: Self-service for tenants (payments, maintenance, documents)
    
    ## Authentication
    - Admin/Staff endpoints use Keycloak Bearer tokens
    - Tenant Portal uses magic link/OTP-based authentication
    
    ## Payment Modes
    - Test mode: Uses Paystack test keys (no real transactions)
    - Live mode: Uses Paystack live keys (real transactions)
    
  version: 1.0.0
  contact:
    name: PropMetrik Support
    email: support@propmetrik.com
servers:
  - url: https://api.propmetrik.com/v1/pm
    description: Production server
  - url: https://staging-api.propmetrik.com/v1/pm
    description: Staging server
  - url: http://localhost:3001/api/pm
    description: Local development
...
paths:
...
  /tenancies:
    get:
      tags:
        - Tenancies
      summary: List tenancies
      operationId: listTenancies
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, expired, terminated]
        - name: propertyId
          in: query
          schema:
            type: string
            format: uuid
        - name: unitId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenancyListResponse'
      security:
        - bearerAuth: []

    post:
      tags:
        - Tenancies
      summary: Create tenancy
      description: Create a new tenancy, generate rent schedule, and optionally send lease for signing
      operationId: createTenancy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenancyRequest'
      responses:
        '201':
          description: Tenancy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenancyDetail'
      security:
        - bearerAuth: []
  /tenancies/{tenancyId}:
    get:
      tags:
        - Tenancies
      summary: Get tenancy details
      operationId: getTenancy
      parameters:
        - $ref: '#/components/parameters/TenancyId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenancyDetail'
      security:
        - bearerAuth: []

    patch:
      tags:
        - Tenancies
      summary: Update tenancy
      operationId: updateTenancy
      parameters:
        - $ref: '#/components/parameters/TenancyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenancyRequest'
      responses:
        '200':
          description: Tenancy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenancyDetail'
      security:
        - bearerAuth: []
...
  /payments/verify/{reference}:
    get:
      tags:
        - Payments
      summary: Verify payment
      description: Verify a Paystack payment transaction
      operationId: verifyPayment
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentVerifyResponse'
      security:
        - bearerAuth: []
...
  /maintenance:
    get:
      tags:
        - Maintenance
      summary: List maintenance requests
      operationId: listMaintenanceRequests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, assigned, in_progress, completed, cancelled]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, emergency]
        - name: propertyId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceListResponse'
      security:
        - bearerAuth: []

    post:
      tags:
        - Maintenance
      summary: Create maintenance request
      operationId: createMaintenanceRequest
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceRequest'
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequest'
      security:
        - bearerAuth: []
  /maintenance/{requestId}:
    get:
      tags:
        - Maintenance
      summary: Get maintenance request
      operationId: getMaintenanceRequest
      parameters:
        - $ref: '#/components/parameters/MaintenanceRequestId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequestDetail'
      security:
        - bearerAuth: []

    patch:
      tags:
        - Maintenance
      summary: Update maintenance request
      description: Update status, assign technician, add notes
      operationId: updateMaintenanceRequest
      parameters:
        - $ref: '#/components/parameters/MaintenanceRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaintenanceRequest'
      responses:
        '200':
          description: Request updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequestDetail'
      security:
        - bearerAuth: []

  # ===== LEASE TEMPLATES =====
...
  /tenant-portal/auth/verify-magic-link:
    post:
      tags:
        - Tenant Portal - Auth
      summary: Verify magic link
      description: Verify magic link token and create session
      operationId: verifyMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSession'
      security: []
...
  /tenant-portal/auth/verify-otp:
    post:
      tags:
        - Tenant Portal - Auth
      summary: Verify OTP
      description: Verify OTP and create session
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - otp
              properties:
                phone:
                  type: string
                otp:
                  type: string
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSession'
      security: []
  /tenant-portal/auth/logout:
    post:
      tags:
        - Tenant Portal - Auth
      summary: Logout
      description: End tenant session
      operationId: tenantLogout
      responses:
        '200':
          description: Logged out
      security:
        - tenantToken: []

  # ===== TENANT PORTAL - PROFILE =====
  /tenant-portal/profile:
    get:
      tags:
        - Tenant Portal - Profile
      summary: Get tenant profile
      operationId: getTenantProfile
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantProfile'
      security:
        - tenantToken: []

    patch:
      tags:
        - Tenant Portal - Profile
      summary: Update tenant profile
      operationId: updateTenantProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantProfile'
      security:
        - tenantToken: []

  # ===== TENANT PORTAL - PAYMENTS =====
...
  /tenant-portal/payments/verify/{reference}:
    get:
      tags:
        - Tenant Portal - Payments
      summary: Verify payment
      operationId: verifyTenantPayment
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentVerifyResponse'
      security:
        - tenantToken: []

  # ===== TENANT PORTAL - MAINTENANCE =====
...
  /tenant-portal/maintenance/{requestId}:
    get:
      tags:
        - Tenant Portal - Maintenance
      summary: Get maintenance request status
      operationId: getTenantMaintenanceStatus
      parameters:
        - $ref: '#/components/parameters/MaintenanceRequestId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequestDetail'
      security:
        - tenantToken: []

  # ===== TENANT PORTAL - DOCUMENTS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT token for admin/staff access
    
    tenantToken:
      type: http
      scheme: bearer
      description: Session token for tenant portal access
  parameters:
...
    TenancyId:
      name: tenancyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MaintenanceRequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TenancyDetail:
      allOf:
        - $ref: '#/components/schemas/Tenancy'
        - type: object
          properties:
            tenant:
              type: object
              properties:
                id:
                  type: string
                fullName:
                  type: string
                email:
                  type: string
                phone:
                  type: string
            property:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                address:
                  type: string
            unit:
              type: object
              properties:
                id:
                  type: string
                unitNumber:
                  type: string
                unitType:
                  type: string
            securityDeposit:
              type: number
            depositPaid:
              type: boolean
            leaseDocumentUrl:
              type: string
    TenancyListResponse:
      type: object
      properties:
        tenancies:
          type: array
          items:
            $ref: '#/components/schemas/TenancyDetail'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateTenancyRequest:
      type: object
      required:
        - propertyId
        - unitId
        - tenantId
        - startDate
        - endDate
        - rentAmount
      properties:
        propertyId:
          type: string
          format: uuid
        unitId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        rentAmount:
          type: number
        rentCurrency:
          type: string
          default: GHS
        paymentFrequency:
          type: string
          enum: [monthly, quarterly, annually]
          default: monthly
        securityDeposit:
          type: number
        generateRentSchedules:
          type: boolean
          default: true
        generateLeaseDocument:
          type: boolean
          default: false
        leaseTemplateId:
          type: string
          format: uuid
    UpdateTenancyRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, expired, terminated]
        rentAmount:
          type: number
        endDate:
          type: string
          format: date

    # Rent Schedule Schemas
...
    PaymentVerifyResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, failed, pending]
        reference:
          type: string
        amount:
          type: number
        currency:
          type: string
        paidAt:
          type: string
          format: date-time
        channel:
          type: string
        message:
          type: string
...
    MaintenanceRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenancyId:
          type: string
          format: uuid
        title:
          type: string
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high, emergency]
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, cancelled]
        description:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
...
    MaintenanceListResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/MaintenanceRequest'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateMaintenanceRequest:
      type: object
      required:
        - tenancyId
        - title
        - category
        - description
      properties:
        tenancyId:
          type: string
          format: uuid
        title:
          type: string
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high, emergency]
          default: medium
        description:
          type: string
        location:
          type: string
        preferredContactMethod:
          type: string
          enum: [phone, email, sms]
        preferredTimeSlot:
          type: string
        images:
          type: array
          items:
            type: string
            format: binary
...
    UpdateTenantProfileRequest:
      type: object
      properties:
        phone:
          type: string
        emergencyContact:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            relationship:
              type: string
...